[31741] - modify the test to use an API wrapper
