+++ dev/com.ibm.ws.microprofile.rest.client_fat/publish/servers/MyRESTClientServer/server.xml
@@ -12,5 +12,7 @@
 
   <!--  Required to read the remote server's port system property -->
   <javaPermission className="java.util.PropertyPermission"  name="*" actions="read" />
+  <javaPermission className="java.lang.RuntimePermission" name="getClassLoader"/>
+  <javaPermission className="java.lang.RuntimePermission" name="createClassLoader"/>
   
 </server>
