+++ dev/com.ibm.ws.microprofile.rest.client_fat/test-applications/MyRestClientBundle/src/io/openliberty/mprestclient/fat/myrestclient/internal/MyRestAPI.java
@@ -1,0 +1,27 @@
+/*******************************************************************************
+ * Copyright (c) 2025 IBM Corporation and others.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-2.0/
+ * 
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package io.openliberty.mprestclient.fat.myrestclient.internal;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.Produces;
+
+import org.eclipse.microprofile.rest.client.inject.RegisterRestClient;
+
+@RegisterRestClient
+@Path("rest")
+public interface MyRestAPI {
+    
+    @GET
+    @Path("greet")
+    @Produces({ "application/json" })
+    public String greet();
+
+}
